fix: resolve numbers loose decimals when read from the database
Closes: #10026

+++ src/driver/aurora-mysql/AuroraMysqlDriver.ts
@@ -636,7 +636,7 @@ export class AuroraMysqlDriver implements Driver {
             value = parseInt(value)
         } else if (columnMetadata.type === Number) {
             // convert to number if number
-            value = !isNaN(+value) ? parseInt(value) : value
+            value = !isNaN(+value) ? parseFloat(value) : value
         }
 
         if (columnMetadata.transformer)
