+++ src/driver/sap/SapDriver.ts
@@ -537,7 +537,7 @@ export class SapDriver implements Driver {
             value = DateUtils.stringToSimpleEnum(value, columnMetadata)
         } else if (columnMetadata.type === Number) {
             // convert to number if number
-            value = !isNaN(+value) ? parseInt(value) : value
+            value = !isNaN(+value) ? parseFloat(value) : value
         }
 
         if (columnMetadata.transformer)
